quantity statistics move to util and made static
