+++ sktime/markov/_base.py
@@ -189,7 +189,6 @@ class BayesianPosterior(Model):
         ----------
         quantity: str
             name of attribute, which will be evaluated on samples
-
         store_samples: bool, optional, default=False
             whether to store the samples (array).
         delimiter : str, optional, default='/'
@@ -204,61 +203,8 @@ class BayesianPosterior(Model):
         statistics : QuantityStatistics
             The statistics
         """
-        from sktime.util import call_member
-        samples = self.samples
-        if delimiter in quantity:
-            qs = quantity.split(delimiter)
-            quantity = qs[-1]
-            for q in qs[:-1]:
-                samples = [call_member(s, q) for s in samples]
-        samples = [call_member(s, quantity, *args, **kwargs) for s in samples]
-        return QuantityStatistics(samples, quantity=quantity, store_samples=store_samples)
-
-
-class QuantityStatistics(Model):
-    """ Container for statistical quantities computed on samples.
-
-    Parameters
-    ----------
-    samples: list of ndarrays
-        the samples
-    store_samples: bool, default=False
-        whether to store the samples (array).
-
-    Attributes
-    ----------
-    mean: array(n)
-        mean along axis=0
-    std: array(n)
-        std deviation along axis=0
-    L : ndarray(shape)
-        element-wise lower bounds
-    R : ndarray(shape)
-        element-wise upper bounds
-    """
-
-    def __init__(self, samples: typing.List[np.ndarray], quantity, confidence=0.95, store_samples=False):
-        super().__init__()
-        self.quantity = quantity
-        # TODO: shall we refer to the original object?
-        # we re-add the (optional) quantity, because the creation of a new array will strip it.
-        unit = getattr(samples[0], 'u', None)
-        if unit is not None:
-            samples = np.array(tuple(x.magnitude for x in samples))
-        else:
-            samples = np.array(samples)
-        if unit is not None:
-            samples *= unit
-        if store_samples:
-            self.samples = samples
-        else:
-            self.samples = np.empty(0) * unit
-        self.mean = samples.mean(axis=0)
-        self.std = samples.std(axis=0)
-        self.L, self.R = confidence_interval(samples, conf=confidence)
-        if unit is not None:
-            self.L *= unit
-            self.R *= unit
+        from sktime.util import QuantityStatistics
+        return QuantityStatistics.gather(self.samples, quantity, store_samples, delimiter, *args, **kwargs)
 
 
 def score_cv(estimator: _MSMBaseEstimator, dtrajs, lagtime, n=10, count_mode="sliding", score_method='VAMP2',
