+++ sktime/markov/msm/__init__.py
@@ -15,4 +15,4 @@ r"""
 from .markov_state_model import MarkovStateModel
 from .maximum_likelihood_msm import MaximumLikelihoodMSM
 from .bayesian_msm import BayesianMSM, BayesianPosterior
-from .._base import QuantityStatistics
+from ...util import QuantityStatistics
